requirements_lock.txt
absl-py==2.4.0
astunparse==1.6.3
beautifulsoup4==4.14.3
certifi==2026.2.25
cffi==2.0.0
charset-normalizer==3.4.7
contourpy==1.3.2
cycler==0.12.1
facenet-pytorch==2.6.0
filelock==3.25.2
flatbuffers==25.12.19
fonttools==4.62.1
fsspec==2026.2.0
gast==0.7.0
gdown==6.0.0
google-pasta==0.2.0
grpcio==1.80.0
h5py==3.14.0
idna==3.11
Jinja2==3.1.6
joblib==1.5.3
keras==3.12.1
kiwisolver==1.5.0
libclang==18.1.1
markdown-it-py==4.0.0
MarkupSafe==3.0.3
matplotlib==3.10.8
mdurl==0.1.2
mediapipe==0.10.33
ml_dtypes==0.5.4
mpmath==1.3.0
namex==0.1.0
networkx==3.4.2
numpy==1.26.4
nvidia-cublas-cu12==12.4.5.8
nvidia-cuda-cupti-cu12==12.4.127
nvidia-cuda-nvrtc-cu12==12.4.127
nvidia-cuda-runtime-cu12==12.4.127
nvidia-cudnn-cu12==9.21.0.82
nvidia-cufft-cu12==11.2.1.3
nvidia-curand-cu12==10.3.5.147
nvidia-cusolver-cu12==11.6.1.9
nvidia-cusparse-cu12==12.3.1.170
nvidia-cusparselt-cu12==0.6.2
nvidia-nccl-cu12==2.21.5
nvidia-nvjitlink-cu12==12.4.127
nvidia-nvtx-cu12==12.4.127
opencv-contrib-python==4.13.0.92
opencv-python==4.13.0.92
opt_einsum==3.4.0
optree==0.19.0
packaging==26.0
pandas==2.3.3
pillow==10.2.0
protobuf==7.34.1
pycparser==3.0
Pygments==2.20.0
pyparsing==3.3.2
PySocks==1.7.1
python-dateutil==2.9.0.post0
pytz==2026.1.post1
requests==2.33.1
rich==15.0.0
scikit-learn==1.7.2
scipy==1.15.3
seaborn==0.13.2
six==1.17.0
sounddevice==0.5.5
soupsieve==2.8.3
sympy==1.13.1
tensorflow==2.21.0
termcolor==3.3.0
threadpoolctl==3.6.0
torch==2.6.0+cu124
torchvision==0.21.0+cu124
tqdm==4.67.3
triton==3.2.0
typing_extensions==4.15.0
tzdata==2026.1
urllib3==2.6.3
wrapt==2.1.2
